+++ synkit/Rule/reactor_rule.py
@@ -1,0 +1,3 @@
+class ReactorRule:
+    def __init__(self) -> None:
+        pass
